add DAC blocking example
